Refactor(handle_hidden.dart): 파일명 변경
get_title.dart에서 handle_hidden.dart로 변경함.
